[1621] Migrate the onboard area to Material UI
Bug: https://github.com/eclipse-sirius/sirius-components/issues/1621
