+++ CHANGELOG.adoc
@@ -20,6 +20,7 @@
 
 - https://github.com/eclipse-sirius/sirius-components/issues/1869[#1869] [tree] Navigate to the first child with the right arrow if a node is expanded.
 Navigate to the parent with the left arrow if a node is collapsed
+- https://github.com/eclipse-sirius/sirius-components/issues/1621[#1621] [project] Migrate the onboard area to Material-UI
 
 
 == v2023.4.0
