+++ packages/core/frontend/sirius-components/src/icons/index.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2022 Obeo.
+ * Copyright (c) 2019, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -15,6 +15,4 @@ export * from './Danger';
 export * from './Delete';
 export * from './Edit';
 export * from './Exit';
-export * from './NewDocument';
-export * from './NewRepresentation';
 export * from './NoIcon';
