--- packages/core/frontend/sirius-components/src/onboarding/AreaContainer.module.css
@@ -1,50 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019, 2022 Obeo.
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Obeo - initial API and implementation
- *******************************************************************************/
-.container {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: min-content min-content 32px;
-  row-gap: 8px;
-  flex-basis: 0;
-}
-
-.titleContainer {
-  display: grid;
-  grid-template-rows: min-content 40px;
-  row-gap: 8px;
-}
-
-.title {
-  font-size: var(--font-size-3);
-  font-weight: var(--font-weight-bold);
-  color: var(--daintree);
-  white-space: nowrap;
-}
-
-.subtitle {
-  font-size: var(--font-size-4);
-  color: var(--daintree-lighten-30);
-}
-
-.content {
-  display: grid;
-  grid-template-rows: 1fr;
-  grid-template-columns: 1fr;
-  row-gap: 16px;
-}
-
-.bannerArea {
-  display: grid;
-  grid-template-rows: 1fr;
-  grid-template-columns: 1fr;
-}
