+++ packages/sirius-web/frontend/sirius-web/src/onboarding/NewDocumentArea.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,8 +10,11 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { Selection } from 'workbench/Workbench.types';
+import { Selection } from '@eclipse-sirius/sirius-components-core';
 
+export interface NewDocumentAreaState {
+  message: string;
+}
 export interface StereotypeDescription {
   id: string;
   label: string;
