+++ packages/sirius-web/frontend/sirius-web/src/onboarding/NewRepresentationArea.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,15 +10,22 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { Selection } from 'workbench/Workbench.types';
+import { Selection } from '@eclipse-sirius/sirius-components-core';
 
-export interface Representation {
-  id: string;
-  label: string;
-  kind: string;
+export interface NewRepresentationAreaState {
+  message: string;
 }
 
-export interface RepresentationAreaProps {
-  representations: Representation[];
+export interface RepresentationDescriptionMetadata {
+  id: string;
+  label: string;
+  defaultName: string;
+}
+
+export interface NewRepresentationAreaProps {
+  editingContextId: string;
+  representationDescriptions: RepresentationDescriptionMetadata[];
+  selection: Selection;
   setSelection: (selection: Selection) => void;
+  readOnly: boolean;
 }
